Add lab4 and configurate CUDA

+++ check_result.py
@@ -7,7 +7,7 @@ def check() -> None:
         matrix1 = np.loadtxt(f"data/1_{i}.txt", dtype=int)
         matrix2 = np.loadtxt(f"data/2_{i}.txt", dtype=int)
         result = np.dot(matrix1, matrix2)
-        cpp_result = np.loadtxt(f"data/result_{i}.txt")
+        cpp_result = np.loadtxt(f"data/result_{i}.txt", dtype=int)
         if np.array_equal(cpp_result, result):
             print(
                 f"Проверка для матрицы {i} успешна: результаты совпадают"
